support distributed in file; refactors
